% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_validate.R
\name{email_validate}
\alias{email_validate}
\title{Validate email address}
\usage{
email_validate(e)
}
\arguments{
\item{e}{email.}
}
\value{
if valid email
}
\description{
Calculate the length of the perimeter of an ellipse.
}
