% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-email.R
\name{email_validate}
\alias{email_validate}
\title{Validate email address}
\usage{
email_validate(x)
}
\arguments{
\item{address}{address for validation}
}
\value{
TRUE if addresss is valid.
}
\description{
Validate email address
}
\examples{
email_validate(edelgadou@gmail.com)
email_validate(delgado_emil@bentley.edu)
}
